refactor: http mod
