+++ server/service/http/etcd_api.go
@@ -17,6 +17,27 @@ type EtcdAPI struct {
 	forwardClient *ForwardClient
 }
 
+type AddMemberRequest struct {
+	MemberAddrs []string `json:"memberAddrs"`
+}
+
+type UpdateMemberRequest struct {
+	OldMemberName string   `json:"oldMemberName"`
+	NewMemberAddr []string `json:"newMemberAddr"`
+}
+
+type RemoveMemberRequest struct {
+	MemberName string `json:"memberName"`
+}
+
+type PromoteLearnerRequest struct {
+	LearnerName string `json:"learnerName"`
+}
+
+type MoveLeaderRequest struct {
+	MemberName string `json:"memberName"`
+}
+
 func NewEtcdAPI(etcdClient *clientv3.Client, forwardClient *ForwardClient) EtcdAPI {
 	return EtcdAPI{
 		etcdClient:    etcdClient,
@@ -24,22 +45,18 @@ func NewEtcdAPI(etcdClient *clientv3.Client, forwardClient *ForwardClient) EtcdA
 	}
 }
 
-type AddMemberRequest struct {
-	MemberAddrs []string `json:"memberAddrs"`
-}
-
 func (a *EtcdAPI) addMember(req *http.Request) apiFuncResult {
 	var addMemberRequest AddMemberRequest
 	err := json.NewDecoder(req.Body).Decode(&addMemberRequest)
 	if err != nil {
 		log.Error("decode request body failed", zap.Error(err))
-		return errResult(ErrParseRequest, fmt.Sprintf("err: %s", err.Error()))
+		return errResult(ErrParseRequest, err.Error())
 	}
 
 	resp, err := a.etcdClient.MemberAdd(req.Context(), addMemberRequest.MemberAddrs)
 	if err != nil {
 		log.Error("member add as learner failed", zap.Error(err))
-		return errResult(ErrAddLearner, fmt.Sprintf("err: %s", err.Error()))
+		return errResult(ErrAddLearner, err.Error())
 	}
 
 	return okResult(resp)
@@ -49,15 +66,10 @@ func (a *EtcdAPI) getMember(req *http.Request) apiFuncResult {
 	resp, err := a.etcdClient.MemberList(req.Context())
 	if err != nil {
 		log.Error("list member failed", zap.Error(err))
-		return errResult(ErrListMembers, fmt.Sprintf("err: %s", err.Error()))
+		return errResult(ErrListMembers, err.Error())
 	}
 
 	return okResult(resp)
-}
-
-type UpdateMemberRequest struct {
-	OldMemberName string   `json:"oldMemberName"`
-	NewMemberAddr []string `json:"newMemberAddr"`
 }
 
 func (a *EtcdAPI) updateMember(req *http.Request) apiFuncResult {
@@ -65,13 +77,13 @@ func (a *EtcdAPI) updateMember(req *http.Request) apiFuncResult {
 	err := json.NewDecoder(req.Body).Decode(&updateMemberRequest)
 	if err != nil {
 		log.Error("decode request body failed", zap.Error(err))
-		return errResult(ErrParseTopology, fmt.Sprintf("err: %s", err.Error()))
+		return errResult(ErrParseTopology, err.Error())
 	}
 
 	memberListResp, err := a.etcdClient.MemberList(req.Context())
 	if err != nil {
 		log.Error("list members failed", zap.Error(err))
-		return errResult(ErrListMembers, fmt.Sprintf("err: %s", err.Error()))
+		return errResult(ErrListMembers, err.Error())
 	}
 
 	for _, member := range memberListResp.Members {
@@ -79,7 +91,7 @@ func (a *EtcdAPI) updateMember(req *http.Request) apiFuncResult {
 			_, err := a.etcdClient.MemberUpdate(req.Context(), member.ID, updateMemberRequest.NewMemberAddr)
 			if err != nil {
 				log.Error("remove learner failed", zap.Error(err))
-				return errResult(ErrRemoveMembers, fmt.Sprintf("err: %s", err.Error()))
+				return errResult(ErrRemoveMembers, err.Error())
 			}
 			return okResult("ok")
 		}
@@ -88,22 +100,18 @@ func (a *EtcdAPI) updateMember(req *http.Request) apiFuncResult {
 	return errResult(ErrGetMember, fmt.Sprintf("member not found, member name: %s", updateMemberRequest.OldMemberName))
 }
 
-type RemoveMemberRequest struct {
-	MemberName string `json:"memberName"`
-}
-
 func (a *EtcdAPI) removeMember(req *http.Request) apiFuncResult {
 	var removeMemberRequest RemoveMemberRequest
 	err := json.NewDecoder(req.Body).Decode(&removeMemberRequest)
 	if err != nil {
 		log.Error("decode request body failed", zap.Error(err))
-		return errResult(ErrParseRequest, fmt.Sprintf("err: %s", err.Error()))
+		return errResult(ErrParseRequest, err.Error())
 	}
 
 	memberListResp, err := a.etcdClient.MemberList(req.Context())
 	if err != nil {
 		log.Error("list members failed", zap.Error(err))
-		return errResult(ErrListMembers, fmt.Sprintf("err: %s", err.Error()))
+		return errResult(ErrListMembers, err.Error())
 	}
 
 	for _, member := range memberListResp.Members {
@@ -111,7 +119,7 @@ func (a *EtcdAPI) removeMember(req *http.Request) apiFuncResult {
 			_, err := a.etcdClient.MemberRemove(req.Context(), member.ID)
 			if err != nil {
 				log.Error("remove learner failed", zap.Error(err))
-				return errResult(ErrRemoveMembers, fmt.Sprintf("err: %s", err.Error()))
+				return errResult(ErrRemoveMembers, err.Error())
 			}
 
 			return okResult("ok")
@@ -121,22 +129,18 @@ func (a *EtcdAPI) removeMember(req *http.Request) apiFuncResult {
 	return errResult(ErrGetMember, fmt.Sprintf("member not found, member name: %s", removeMemberRequest.MemberName))
 }
 
-type PromoteLearnerRequest struct {
-	LearnerName string `json:"learnerName"`
-}
-
 func (a *EtcdAPI) promoteLearner(req *http.Request) apiFuncResult {
 	var promoteLearnerRequest PromoteLearnerRequest
 	err := json.NewDecoder(req.Body).Decode(&promoteLearnerRequest)
 	if err != nil {
 		log.Error("decode request body failed", zap.Error(err))
-		return errResult(ErrParseRequest, fmt.Sprintf("err: %s", err.Error()))
+		return errResult(ErrParseRequest, err.Error())
 	}
 
 	memberListResp, err := a.etcdClient.MemberList(req.Context())
 	if err != nil {
 		log.Error("list members failed", zap.Error(err))
-		return errResult(ErrListMembers, fmt.Sprintf("err: %s", err.Error()))
+		return errResult(ErrListMembers, err.Error())
 	}
 
 	for _, member := range memberListResp.Members {
@@ -144,7 +148,7 @@ func (a *EtcdAPI) promoteLearner(req *http.Request) apiFuncResult {
 			_, err := a.etcdClient.MemberPromote(req.Context(), member.ID)
 			if err != nil {
 				log.Error("remove learner failed", zap.Error(err))
-				return errResult(ErrRemoveMembers, fmt.Sprintf("err: %s", err.Error()))
+				return errResult(ErrRemoveMembers, err.Error())
 			}
 			return okResult("ok")
 		}
@@ -153,22 +157,18 @@ func (a *EtcdAPI) promoteLearner(req *http.Request) apiFuncResult {
 	return errResult(ErrGetMember, fmt.Sprintf("learner not found, learner name: %s", promoteLearnerRequest.LearnerName))
 }
 
-type MoveLeaderRequest struct {
-	MemberName string `json:"memberName"`
-}
-
 func (a *EtcdAPI) moveLeader(req *http.Request) apiFuncResult {
 	var moveLeaderRequest MoveLeaderRequest
 	err := json.NewDecoder(req.Body).Decode(&moveLeaderRequest)
 	if err != nil {
 		log.Error("decode request body failed", zap.Error(err))
-		return errResult(ErrParseRequest, fmt.Sprintf("err: %s", err.Error()))
+		return errResult(ErrParseRequest, err.Error())
 	}
 
 	memberListResp, err := a.etcdClient.MemberList(req.Context())
 	if err != nil {
 		log.Error("list members failed", zap.Error(err))
-		return errResult(ErrListMembers, fmt.Sprintf("err: %s", err.Error()))
+		return errResult(ErrListMembers, err.Error())
 	}
 
 	for _, member := range memberListResp.Members {
@@ -176,7 +176,7 @@ func (a *EtcdAPI) moveLeader(req *http.Request) apiFuncResult {
 			moveLeaderResp, err := a.etcdClient.MoveLeader(req.Context(), member.ID)
 			if err != nil {
 				log.Error("remove learner failed", zap.Error(err))
-				return errResult(ErrRemoveMembers, fmt.Sprintf("err: %s", err.Error()))
+				return errResult(ErrRemoveMembers, err.Error())
 			}
 			log.Info("move leader", zap.String("moveLeaderResp", fmt.Sprintf("%v", moveLeaderResp)))
 			return okResult("ok")
